+++ .env.example
@@ -1,0 +1,3 @@
+NODE_ENV=dev
+BACK_URL=http://localhost:127.0.0.1/
+YAMAP_API_KEY=your-individual-key
