+++ package.json
@@ -29,8 +29,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^9.0.2",
-    "react-router-dom": "^6.20.0",
-    "react-yandex-maps": "^4.6.0"
+    "react-router-dom": "^6.20.0"
   },
   "devDependencies": {
     "@storybook/addon-essentials": "^7.6.2",
