+++ src/components/App/App.test.ts
@@ -1,4 +1,3 @@
-// sum.test.js
 import { expect, test } from 'vitest';
 
 test('true test', () => {
