+++ src/components/pages/MainPage/MapComponent/MapComponent.scss
@@ -12,13 +12,18 @@
     }
   }
 
-&__group {
-  display: flex;
-  justify-content: space-between;
-
-  &-switch {
+  &__group {
     display: flex;
-    align-items: center;
+    justify-content: space-between;
+
+    &-switch {
+      display: flex;
+      align-items: center;
+    }
   }
-}
+
+  &__container {
+    width: 100%;
+    height: 40vh;
+  }
 }
