+++ src/components/pages/MainPage/SearchClinic/SearchClinic.scss
@@ -1,4 +1,5 @@
 .search-clinic {
+  min-width: 800px;
   margin: 0 auto;
   width: 90%;
 }
