+++ src/components/pages/ReduxPage/ReduxPage.scss
@@ -1,4 +1,4 @@
-.main {
+.redux-page {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -16,10 +16,12 @@
   padding: 1.5em;
   will-change: filter;
   transition: filter 300ms;
+
+  &:hover {
+    filter: drop-shadow(0 0 2em #646cffaa);
+  }
 }
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
+
 .logo.react:hover {
   filter: drop-shadow(0 0 2em #61dafbaa);
 }
@@ -41,24 +43,24 @@
 
 .card {
   padding: 2em;
+
+  &__content {
+    margin-bottom: 20px;
+
+    display: flex;
+    gap: 10px;
+    justify-content: center;
+    align-items: center;
+
+    font-size: 20px;
+  }
+
+  &__btn {
+    min-width: 40px;
+    min-height: 30px;
+  }
 }
 
 .read-the-docs {
   color: #888;
-}
-
-.card__content {
-  margin-bottom: 20px;
-
-  display: flex;
-  gap: 10px;
-  justify-content: center;
-  align-items: center;
-
-  font-size: 20px;
-}
-
-.card__btn {
-  min-width: 40px;
-  min-height: 30px;
 }
